refactor: use controller pattern in rest api, fix permissions
